Add ItemGrid as global component to use across item screens
