+++ src/components/slots/Inventory.vue
@@ -1,31 +1,12 @@
 <template>
-  <div
-    class="inventory_slot">
-    <div
-      v-for="(n, i) in (0, 24)"
-      :key="i"
-    >
-      <div
-        v-if="slotHasItem(i)"
-        :style="{
-          backgroundImage: 'url(' + getBgUrl(i) + ')',
-          backgroundPosition: `left -${(getItem(i).column * 32)}px top -${(getItem(i).row * 32)}px`
-        }"
-        class="slot inventorySlot"
-        @click.right="rightClick($event, i)" />
-      <div
-        v-else
-        class="slot inventorySlot">
-        <!-- Empty slot -->
-      </div>
-    </div>
-  </div>
+  <item-grid
+    :images="game.map.images"
+    :items="items"
+    :slots="24"
+    screen="inventory" />
 </template>
 
 <script>
-import UI from 'shared/ui';
-import bus from '../../core/utilities/bus';
-
 export default {
   props: {
     game: {
@@ -39,98 +20,6 @@ export default {
     },
     items() {
       return this.game.player.inventory;
-    },
-  },
-  methods: {
-    /**
-     * Check to see if this slot is available
-     *
-     * @param {integer} slotNumber The index of the slot in inventory
-     * @return {boolean}
-     */
-    slotHasItem(slotNumber) {
-      const itemFound = this.items.find(s => s.slot === slotNumber);
-      if (!itemFound) {
-        return false;
-      }
-
-      if (itemFound) {
-        return true;
-      }
-
-      return false;
-    },
-    /**
-     * Get the item's column of a certain slot in the inventory
-     *
-     * @param {integer} slotNumber The slot index in the inventory
-     * @return {integer|boolean}
-     */
-    getItem(slotNumber) {
-      const getItem = this.getItemFromSlot(slotNumber);
-
-      if (getItem) {
-        return {
-          column: getItem.graphics.column,
-          row: getItem.graphics.row,
-        };
-      }
-
-      return false;
-    },
-
-    /**
-     * Right-click brings up context-menu
-     *
-     * @param {event} event The mouse-click event
-     */
-    rightClick(event, index) {
-      const coordinates = UI.getViewportCoordinates(event);
-
-      const data = {
-        event,
-        coordinates,
-        slot: index,
-        target: event.target,
-      };
-
-      event.preventDefault();
-
-      bus.$emit('PLAYER:MENU', data);
-    },
-    /**
-     * Gets the item from current slot
-     *
-     * @param {integer} slotNumber The current slot number
-     * @returns {object}
-     */
-    getItemFromSlot(slotNumber) {
-      return this.items.find(s => s.slot === slotNumber);
-    },
-    /**
-     * Get the correct background URL to show in inventory
-     *
-     * @param {string} item The item type in slot
-     * @returns {string}
-     */
-    getBgUrl(slotNumber) {
-      const getItem = this.getItemFromSlot(slotNumber);
-
-      if (!this.images) {
-        return false;
-      }
-
-      switch (getItem.graphics.tileset) {
-        case 'general':
-          return this.images.generalImage.src;
-        case 'jewelry':
-          return this.images.jewelryImage.src;
-        case 'armor':
-          return this.images.armorImage.src;
-        default:
-        case 'weapons':
-          return this.images.weaponsImage.src;
-      }
     },
   },
 };
